Move ICloseable interface; add IOUtils Close method implementations
