+++ src/Lucene.Net/Util/IOUtils.cs
@@ -58,8 +58,66 @@ namespace Lucene.Net.Util
         public static readonly string UTF_8 = "UTF-8";
 
         /// <summary>
-        /// <para>Disposes all given <c>IDisposable</c>s, suppressing all thrown exceptions. Some of the <c>IDisposable</c>s
-        /// may be <c>null</c>, they are ignored. After everything is disposed, method either throws <paramref name="priorException"/>,
+        /// Closes all given <see cref="ICloseable"/>s.  Some of the
+        /// <see cref="ICloseable"/>s may be <c>null</c>; they are
+        /// ignored.  After everything is closed, the method either
+        /// throws the first exception it hit while closing, or
+        /// completes normally if there were no exceptions.
+        /// </summary>
+        /// <param name="objects">Objects to call <see cref="ICloseable.Close()"/> on</param>
+        public static void Close(params ICloseable[] objects)
+        {
+            Exception th = null;
+
+            foreach (ICloseable @object in objects)
+            {
+                try
+                {
+                    @object?.Close();
+                }
+                catch (Exception t) when (t.IsThrowable())
+                {
+                    AddSuppressed(th, t);
+                    if (th is null)
+                    {
+                        th = t;
+                    }
+                }
+            }
+
+            ReThrow(th);
+        }
+
+        /// <summary>
+        /// Closes all given <see cref="ICloseable"/>s.
+        /// </summary>
+        /// <seealso cref="Close(ICloseable[])"/>
+        public static void Close(IEnumerable<ICloseable> objects)
+        {
+            Exception th = null;
+
+            foreach (ICloseable @object in objects)
+            {
+                try
+                {
+                    @object?.Close();
+                }
+                catch (Exception t) when (t.IsThrowable())
+                {
+                    AddSuppressed(th, t);
+                    if (th is null)
+                    {
+                        th = t;
+                    }
+                }
+            }
+
+            ReThrow(th);
+        }
+
+        /// <summary>
+        /// <para>Closes all given <see cref="ICloseable">ICloseable</see>s, suppressing all thrown exceptions. Some of the <see cref="ICloseable">ICloseable</see>s
+        /// may be <c>null</c>, they are ignored. After everything is closed, method either throws <paramref name="priorException"/>,
         /// if one is supplied, or the first of suppressed exceptions, or completes normally.</para>
         /// <para>Sample usage:
         /// <code>
@@ -82,10 +140,10 @@ namespace Lucene.Net.Util
         /// </para>
         /// </summary>
         /// <param name="priorException">  <c>null</c> or an exception that will be rethrown after method completion. </param>
-        /// <param name="objects">         Objects to call <see cref="IDisposable.Dispose()"/> on. </param>
+        /// <param name="objects">         Objects to call <see cref="ICloseable.Close()"/> on. </param>
         public static void CloseWhileHandlingException(Exception priorException, params ICloseable[] objects)
         {
-            CloseWhileHandlingException(priorException, objects);
+            throw new NotImplementedException("TODO: determine if these are still needed, they don't exist upstream");
         }
 
         /// <summary>
@@ -94,29 +152,7 @@ namespace Lucene.Net.Util
         /// <seealso cref="CloseWhileHandlingException(Exception, ICloseable[])"/>
         public static void CloseWhileHandlingException(Exception priorException, IEnumerable<ICloseable> objects)
         {
-            CloseWhileHandlingException(priorException, objects);
-        }
-
-        /// <summary>
-        /// Closes all given <see cref="ICloseable"/>s.  Some of the
-        /// <see cref="ICloseable"/>s may be <c>null</c>; they are
-        /// ignored.  After everything is closed, the method either
-        /// throws the first exception it hit while closing, or
-        /// completes normally if there were no exceptions.
-        /// </summary>
-        /// <param name="objects">Objects to call <see cref="ICloseable.Close()"/> on</param>
-        public static void Close(params ICloseable[] objects)
-        {
-            Close(objects);
-        }
-
-        /// <summary>
-        /// Closes all given <see cref="ICloseable"/>s.
-        /// </summary>
-        /// <seealso cref="Close(ICloseable[])"/>
-        public static void Close(IEnumerable<ICloseable> objects)
-        {
-            Close(objects);
+            throw new NotImplementedException("TODO: determine if these are still needed, they don't exist upstream");
         }
 
         /// <summary>
@@ -126,7 +162,25 @@ namespace Lucene.Net.Util
         /// <param name="objects">Objects to call <see cref="ICloseable.Close()"/> on</param>
         public static void CloseWhileHandlingException(params ICloseable[] objects)
         {
-            CloseWhileHandlingException(objects);
+            Exception th = null;
+
+            foreach (ICloseable @object in objects)
+            {
+                try
+                {
+                    @object?.Close();
+                }
+                catch (Exception t) when (t.IsThrowable())
+                {
+                    AddSuppressed(th, t);
+                    if (th is null)
+                    {
+                        th = t;
+                    }
+                }
+            }
+
+            ReThrow(th);
         }
 
         /// <summary>
@@ -136,7 +190,25 @@ namespace Lucene.Net.Util
         /// <seealso cref="CloseWhileHandlingException(ICloseable[])"/>
         public static void CloseWhileHandlingException(IEnumerable<ICloseable> objects)
         {
-            CloseWhileHandlingException(objects);
+            Exception th = null;
+
+            foreach (ICloseable @object in objects)
+            {
+                try
+                {
+                    @object?.Close();
+                }
+                catch (Exception t) when (t.IsThrowable())
+                {
+                    AddSuppressed(th, t);
+                    if (th is null)
+                    {
+                        th = t;
+                    }
+                }
+            }
+
+            ReThrow(th);
         }
